+++ libs/pavexc/src/compiler/analyses/components/db/mod.rs
@@ -1,314 +1,39 @@
-use std::borrow::Cow;
-use std::collections::BTreeMap;
-
-use ahash::{HashMap, HashMapExt, HashSet};
-use guppy::graph::PackageGraph;
-use indexmap::IndexSet;
-use miette::NamedSource;
-use rustdoc_types::ItemEnum;
-use syn::spanned::Spanned;
-
-use pavex_bp_schema::{CloningStrategy, Lifecycle, Lint, LintSetting};
-
+use crate::compiler::analyses::components::component::Component;
+use crate::compiler::analyses::components::{
+    unregistered::UnregisteredComponent, ConsumptionMode, HydratedComponent, InsertTransformer,
+    SourceId,
+};
 use crate::compiler::analyses::computations::{ComputationDb, ComputationId};
+use crate::compiler::analyses::framework_items::FrameworkItemDb;
 use crate::compiler::analyses::into_error::register_error_new_transformer;
 use crate::compiler::analyses::user_components::{
     ScopeGraph, ScopeId, UserComponent, UserComponentDb, UserComponentId,
 };
 use crate::compiler::component::{
-    Constructor, ConstructorValidationError, ErrorHandler, ErrorHandlerValidationError,
-    ErrorObserver, ErrorObserverValidationError, RequestHandler, RequestHandlerValidationError,
-    WrappingMiddleware, WrappingMiddlewareValidationError,
+    Constructor, ConstructorValidationError, ErrorHandler, ErrorObserver, RequestHandler,
+    WrappingMiddleware,
 };
 use crate::compiler::computation::{Computation, MatchResult};
 use crate::compiler::interner::Interner;
-use crate::compiler::resolvers::CallableResolutionError;
-use crate::compiler::traits::{assert_trait_is_implemented, MissingTraitImplementationError};
+use crate::compiler::traits::assert_trait_is_implemented;
 use crate::compiler::utils::{
     get_err_variant, get_ok_variant, process_framework_callable_path, process_framework_path,
-};
-use crate::diagnostic;
-use crate::diagnostic::{
-    convert_proc_macro_span, convert_rustdoc_span, AnnotatedSnippet, CallableType,
-    CompilerDiagnostic, LocationExt, SourceSpanExt,
 };
 use crate::language::{
     Callable, Lifetime, PathType, ResolvedPath, ResolvedPathQualifiedSelf, ResolvedPathSegment,
     ResolvedType, TypeReference,
 };
 use crate::rustdoc::CrateCollection;
-use crate::utils::comma_separated_list;
+use ahash::{HashMap, HashMapExt, HashSet};
+use guppy::graph::PackageGraph;
+use indexmap::IndexSet;
+use pavex_bp_schema::{CloningStrategy, Lifecycle, Lint, LintSetting};
+use std::borrow::Cow;
+use std::collections::BTreeMap;
 
-use super::framework_items::FrameworkItemDb;
-
-/// Describe the relationship between this component and one of its input parameters with
-/// respect to Rust's ownership semantics.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub(crate) enum ConsumptionMode {
-    /// The component takes the input by value, consuming it (e.g. `fn f(t: MyStruct)`).
-    Move,
-    /// The component takes a shared borrow of the input (e.g. `fn f(t: &MyStruct)`).
-    SharedBorrow,
-}
-
-/// When should the transformer node be inserted in the graph?
-#[derive(Debug, Clone, Copy, Hash, Ord, PartialOrd, Eq, PartialEq)]
-pub(crate) enum InsertTransformer {
-    /// Always insert the transformer node if the transformed component appears in the graph.
-    Eagerly,
-    /// Don't automatically insert the transformer node. Instead, the compiler
-    /// will manually insert it when it is needed.
-    ///
-    /// This is primarily used for cloning nodes.
-    Lazily,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub(crate) enum Component {
-    RequestHandler {
-        user_component_id: UserComponentId,
-    },
-    WrappingMiddleware {
-        source_id: SourceId,
-    },
-    ErrorHandler {
-        source_id: SourceId,
-    },
-    ErrorObserver {
-        user_component_id: UserComponentId,
-    },
-    Constructor {
-        source_id: SourceId,
-    },
-    Transformer {
-        computation_id: ComputationId,
-        transformed_component_id: ComponentId,
-        transformation_mode: ConsumptionMode,
-        scope_id: ScopeId,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-/// All the information needed to register a component.
-/// Some of this information will not be attached directly to the final [`Component`],
-/// but it'll end up being tracked out of band in [`ComponentDb`] (e.g. the lifecycle in
-/// `ComponentDb::id2lifecycle`).
-pub(crate) enum UnregisteredComponent {
-    RequestHandler {
-        user_component_id: UserComponentId,
-    },
-    UserWrappingMiddleware {
-        user_component_id: UserComponentId,
-    },
-    SyntheticWrappingMiddleware {
-        computation_id: ComputationId,
-        scope_id: ScopeId,
-    },
-    ErrorHandler {
-        source_id: SourceId,
-        fallible_component_id: ComponentId,
-        error_handler: ErrorHandler,
-    },
-    ErrorObserver {
-        user_component_id: UserComponentId,
-        error_input_index: usize,
-    },
-    UserConstructor {
-        user_component_id: UserComponentId,
-    },
-    SyntheticConstructor {
-        lifecycle: Lifecycle,
-        computation_id: ComputationId,
-        scope_id: ScopeId,
-        cloning_strategy: CloningStrategy,
-        /// Synthetic constructors can be built by "deriving" user-registered constructors.
-        /// For example, by binding unassigned generic parameters or by extracting the `Ok` variant
-        /// from the output of fallible constructors.
-        ///
-        /// If that's the case,
-        /// this field should be populated with the id of the "source" constructor.
-        derived_from: Option<ComponentId>,
-    },
-    Transformer {
-        computation_id: ComputationId,
-        transformed_component_id: ComponentId,
-        transformation_mode: ConsumptionMode,
-        scope_id: ScopeId,
-        when_to_insert: InsertTransformer,
-    },
-}
-
-impl UnregisteredComponent {
-    /// Convert this unregistered component into a registered component **without interning it**.
-    pub fn component(&self) -> Component {
-        match self {
-            UnregisteredComponent::RequestHandler { user_component_id } => {
-                Component::RequestHandler {
-                    user_component_id: user_component_id.to_owned(),
-                }
-            }
-            UnregisteredComponent::UserWrappingMiddleware { user_component_id } => {
-                Component::WrappingMiddleware {
-                    source_id: SourceId::UserComponentId(user_component_id.to_owned()),
-                }
-            }
-            UnregisteredComponent::SyntheticWrappingMiddleware {
-                computation_id,
-                scope_id,
-            } => Component::WrappingMiddleware {
-                source_id: SourceId::ComputationId(computation_id.to_owned(), *scope_id),
-            },
-            UnregisteredComponent::ErrorHandler { source_id, .. } => Component::ErrorHandler {
-                source_id: source_id.to_owned(),
-            },
-            UnregisteredComponent::ErrorObserver {
-                user_component_id, ..
-            } => Component::ErrorObserver {
-                user_component_id: user_component_id.to_owned(),
-            },
-            UnregisteredComponent::UserConstructor { user_component_id } => {
-                Component::Constructor {
-                    source_id: SourceId::UserComponentId(user_component_id.to_owned()),
-                }
-            }
-            UnregisteredComponent::SyntheticConstructor {
-                computation_id,
-                scope_id,
-                ..
-            } => Component::Constructor {
-                source_id: SourceId::ComputationId(computation_id.to_owned(), *scope_id),
-            },
-            UnregisteredComponent::Transformer {
-                computation_id,
-                transformed_component_id,
-                transformation_mode,
-                scope_id,
-                ..
-            } => Component::Transformer {
-                computation_id: computation_id.to_owned(),
-                transformed_component_id: transformed_component_id.to_owned(),
-                transformation_mode: transformation_mode.to_owned(),
-                scope_id: scope_id.to_owned(),
-            },
-        }
-    }
-
-    fn lifecycle(&self, component_db: &ComponentDb) -> Lifecycle {
-        use UnregisteredComponent::*;
-        match &self {
-            UserWrappingMiddleware { .. }
-            | SyntheticWrappingMiddleware { .. }
-            | RequestHandler { .. } => Lifecycle::RequestScoped,
-            ErrorObserver { .. } => Lifecycle::Transient,
-            SyntheticConstructor { lifecycle, .. } => lifecycle.to_owned(),
-            ErrorHandler {
-                fallible_component_id: id,
-                ..
-            }
-            | Transformer {
-                transformed_component_id: id,
-                ..
-            } => component_db.lifecycle(*id),
-            UserConstructor {
-                user_component_id, ..
-            } => component_db
-                .user_component_db
-                .get_lifecycle(*user_component_id)
-                .to_owned(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub(crate) enum SourceId {
-    ComputationId(ComputationId, ScopeId),
-    UserComponentId(UserComponentId),
-}
-
-impl From<UserComponentId> for SourceId {
-    fn from(value: UserComponentId) -> Self {
-        Self::UserComponentId(value)
-    }
-}
+pub(crate) mod diagnostics;
 
 pub(crate) type ComponentId = la_arena::Idx<Component>;
-
-/// A transformation that, given a set of inputs, **constructs** a new type.
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub(crate) enum HydratedComponent<'a> {
-    Constructor(Constructor<'a>),
-    RequestHandler(RequestHandler<'a>),
-    WrappingMiddleware(WrappingMiddleware<'a>),
-    ErrorHandler(Cow<'a, ErrorHandler>),
-    Transformer(Computation<'a>),
-    ErrorObserver(ErrorObserver<'a>),
-}
-
-impl<'a> HydratedComponent<'a> {
-    pub(crate) fn input_types(&self) -> Cow<[ResolvedType]> {
-        match self {
-            HydratedComponent::Constructor(c) => c.input_types(),
-            HydratedComponent::RequestHandler(r) => Cow::Borrowed(r.input_types()),
-            HydratedComponent::ErrorHandler(e) => Cow::Borrowed(e.input_types()),
-            HydratedComponent::Transformer(c) => c.input_types(),
-            HydratedComponent::WrappingMiddleware(c) => Cow::Borrowed(c.input_types()),
-            HydratedComponent::ErrorObserver(eo) => Cow::Borrowed(eo.input_types()),
-        }
-    }
-
-    pub(crate) fn output_type(&self) -> Option<&ResolvedType> {
-        match self {
-            HydratedComponent::Constructor(c) => Some(c.output_type()),
-            HydratedComponent::RequestHandler(r) => Some(r.output_type()),
-            HydratedComponent::ErrorHandler(e) => Some(e.output_type()),
-            HydratedComponent::WrappingMiddleware(e) => Some(e.output_type()),
-            // TODO: we are not enforcing that the output type of a transformer is not
-            //  the unit type. In particular, you can successfully register a `Result<T, ()>`
-            //  type, which will result into a `MatchResult` with output `()` for the error.
-            HydratedComponent::Transformer(c) => c.output_type(),
-            HydratedComponent::ErrorObserver(_) => None,
-        }
-    }
-
-    pub(crate) fn is_fallible(&self) -> bool {
-        self.output_type().map_or(false, |t| t.is_result())
-    }
-
-    /// Returns a [`Computation`] that matches the transformation carried out by this component.
-    pub(crate) fn computation(&self) -> Computation<'a> {
-        match self {
-            HydratedComponent::Constructor(c) => c.0.clone(),
-            HydratedComponent::RequestHandler(r) => r.callable.clone().into(),
-            HydratedComponent::WrappingMiddleware(w) => w.callable.clone().into(),
-            HydratedComponent::ErrorHandler(e) => e.callable.clone().into(),
-            HydratedComponent::Transformer(t) => t.clone(),
-            HydratedComponent::ErrorObserver(eo) => eo.callable.clone().into(),
-        }
-    }
-
-    pub(crate) fn into_owned(self) -> HydratedComponent<'static> {
-        match self {
-            HydratedComponent::Constructor(c) => HydratedComponent::Constructor(c.into_owned()),
-            HydratedComponent::RequestHandler(r) => {
-                HydratedComponent::RequestHandler(r.into_owned())
-            }
-            HydratedComponent::WrappingMiddleware(w) => {
-                HydratedComponent::WrappingMiddleware(w.into_owned())
-            }
-            HydratedComponent::ErrorHandler(e) => {
-                HydratedComponent::ErrorHandler(Cow::Owned(e.into_owned()))
-            }
-            HydratedComponent::Transformer(t) => HydratedComponent::Transformer(t.into_owned()),
-            HydratedComponent::ErrorObserver(eo) => {
-                HydratedComponent::ErrorObserver(eo.into_owned())
-            }
-        }
-    }
-
-    pub(crate) fn is_error_handler(&self) -> bool {
-        matches!(self, HydratedComponent::ErrorHandler(_))
-    }
-}
 
 #[derive(Debug)]
 pub(crate) struct ComponentDb {
@@ -568,7 +293,7 @@ impl ComponentDb {
             .insert(id, unregistered_component.lifecycle(self));
 
         {
-            use UnregisteredComponent::*;
+            use crate::compiler::analyses::components::UnregisteredComponent::*;
             match unregistered_component {
                 ErrorHandler {
                     fallible_component_id,
@@ -905,7 +630,7 @@ impl ComponentDb {
             .user_component_db
             .iter()
             .filter_map(|(id, c)| {
-                use UserComponent::*;
+                use crate::compiler::analyses::user_components::UserComponent::*;
                 match c {
                     ErrorHandler {
                         fallible_callable_identifiers_id,
@@ -1057,7 +782,7 @@ impl ComponentDb {
             .interner
             .iter()
             .filter_map(|(id, c)| {
-                use Component::*;
+                use crate::compiler::analyses::components::component::Component::*;
 
                 match c {
                     RequestHandler { user_component_id }
@@ -1706,856 +1431,5 @@ impl std::ops::Index<ComponentId> for ComponentDb {
 
     fn index(&self, index: ComponentId) -> &Self::Output {
         &self.interner[index]
-    }
-}
-
-/// Utility functions to produce diagnostics.
-impl ComponentDb {
-    fn invalid_constructor(
-        e: ConstructorValidationError,
-        user_component_id: UserComponentId,
-        user_component_db: &UserComponentDb,
-        computation_db: &ComputationDb,
-        package_graph: &PackageGraph,
-        krate_collection: &CrateCollection,
-        diagnostics: &mut Vec<miette::Error>,
-    ) {
-        let location = user_component_db.get_location(user_component_id);
-        let source = match location.source_file(package_graph) {
-            Ok(s) => s,
-            Err(e) => {
-                diagnostics.push(e.into());
-                return;
-            }
-        };
-        let label = diagnostic::get_f_macro_invocation_span(&source, location)
-            .map(|s| s.labeled("The constructor was registered here".into()));
-        let diagnostic = match e {
-            ConstructorValidationError::CannotFalliblyReturnTheUnitType
-            | ConstructorValidationError::CannotReturnTheUnitType => {
-                CompilerDiagnostic::builder(source, e)
-                    .optional_label(label)
-                    .build()
-            }
-            ConstructorValidationError::UnderconstrainedGenericParameters { ref parameters } => {
-                fn get_definition_span(
-                    callable: &Callable,
-                    free_parameters: &IndexSet<String>,
-                    krate_collection: &CrateCollection,
-                    package_graph: &PackageGraph,
-                ) -> Option<AnnotatedSnippet> {
-                    let global_item_id = callable.source_coordinates.as_ref()?;
-                    let item = krate_collection.get_type_by_global_type_id(global_item_id);
-                    let definition_span = item.span.as_ref()?;
-                    let source_contents = diagnostic::read_source_file(
-                        &definition_span.filename,
-                        &package_graph.workspace(),
-                    )
-                    .ok()?;
-                    let span = convert_rustdoc_span(&source_contents, definition_span.to_owned());
-                    let span_contents =
-                        source_contents[span.offset()..(span.offset() + span.len())].to_string();
-                    let (generic_params, output) = match &item.inner {
-                        ItemEnum::Function(_) => {
-                            if let Ok(item) = syn::parse_str::<syn::ItemFn>(&span_contents) {
-                                (item.sig.generics.params, item.sig.output)
-                            } else if let Ok(item) =
-                                syn::parse_str::<syn::ImplItemFn>(&span_contents)
-                            {
-                                (item.sig.generics.params, item.sig.output)
-                            } else {
-                                panic!("Could not parse as a function or method:\n{span_contents}")
-                            }
-                        }
-                        _ => unreachable!(),
-                    };
-
-                    let mut labels = vec![];
-                    let subject_verb = if generic_params.len() == 1 {
-                        "it is"
-                    } else {
-                        "they are"
-                    };
-                    for param in generic_params {
-                        if let syn::GenericParam::Type(ty) = param {
-                            if free_parameters.contains(ty.ident.to_string().as_str()) {
-                                labels.push(
-                                    convert_proc_macro_span(&span_contents, ty.span())
-                                        .labeled("I can't infer this..".into()),
-                                );
-                            }
-                        }
-                    }
-                    let output_span = if let syn::ReturnType::Type(_, output_type) = &output {
-                        output_type.span()
-                    } else {
-                        output.span()
-                    };
-                    labels.push(
-                        convert_proc_macro_span(&span_contents, output_span)
-                            .labeled(format!("..because {subject_verb} not used here")),
-                    );
-                    let source_path = definition_span.filename.to_str().unwrap();
-                    Some(AnnotatedSnippet::new_with_labels(
-                        NamedSource::new(source_path, span_contents),
-                        labels,
-                    ))
-                }
-
-                let callable = &computation_db[user_component_id];
-                let definition_snippet =
-                    get_definition_span(callable, parameters, krate_collection, package_graph);
-                let subject_verb = if parameters.len() == 1 {
-                    "it is"
-                } else {
-                    "they are"
-                };
-                let free_parameters = if parameters.len() == 1 {
-                    format!("`{}`", &parameters[0])
-                } else {
-                    let mut buffer = String::new();
-                    comma_separated_list(
-                        &mut buffer,
-                        parameters.iter(),
-                        |p| format!("`{}`", p),
-                        "and",
-                    )
-                    .unwrap();
-                    buffer
-                };
-                let error = anyhow::anyhow!(e)
-                    .context(
-                        format!(
-                            "I am not smart enough to figure out the concrete type for all the generic parameters in `{}`.\n\
-                            I can only infer the type of an unassigned generic parameter if it appears in the output type returned by the constructor. This is \
-                            not the case for {free_parameters}, since {subject_verb} only used by the input parameters.",
-                            callable.path));
-                CompilerDiagnostic::builder(source, error)
-                    .optional_label(label)
-                    .optional_additional_annotated_snippet(definition_snippet)
-                    .help(
-                        "Specify the concrete type(s) for the problematic \
-                        generic parameter(s) when registering the constructor against the blueprint: \n\
-                        |  bp.constructor(\n\
-                        |    f!(my_crate::my_constructor::<ConcreteType>), \n\
-                        |    ..\n\
-                        |  )".into())
-                    // ^ TODO: add a proper code snippet here, using the actual function that needs
-                    //    to be amended instead of a made signature
-                    .build()
-            }
-            ConstructorValidationError::NakedGenericOutputType {
-                ref naked_parameter,
-            } => {
-                fn get_definition_span(
-                    callable: &Callable,
-                    krate_collection: &CrateCollection,
-                    package_graph: &PackageGraph,
-                ) -> Option<AnnotatedSnippet> {
-                    let global_item_id = callable.source_coordinates.as_ref()?;
-                    let item = krate_collection.get_type_by_global_type_id(global_item_id);
-                    let definition_span = item.span.as_ref()?;
-                    let source_contents = diagnostic::read_source_file(
-                        &definition_span.filename,
-                        &package_graph.workspace(),
-                    )
-                    .ok()?;
-                    let span = convert_rustdoc_span(&source_contents, definition_span.to_owned());
-                    let span_contents =
-                        source_contents[span.offset()..(span.offset() + span.len())].to_string();
-                    let output = match &item.inner {
-                        ItemEnum::Function(_) => {
-                            if let Ok(item) = syn::parse_str::<syn::ItemFn>(&span_contents) {
-                                item.sig.output
-                            } else if let Ok(item) =
-                                syn::parse_str::<syn::ImplItemFn>(&span_contents)
-                            {
-                                item.sig.output
-                            } else {
-                                panic!("Could not parse as a function or method:\n{span_contents}")
-                            }
-                        }
-                        _ => unreachable!(),
-                    };
-
-                    let output_span = if let syn::ReturnType::Type(_, output_type) = &output {
-                        output_type.span()
-                    } else {
-                        output.span()
-                    };
-                    let label = convert_proc_macro_span(&span_contents, output_span)
-                        .labeled("The invalid output type".to_string());
-                    let source_path = definition_span.filename.to_str().unwrap();
-                    Some(AnnotatedSnippet::new(
-                        NamedSource::new(source_path, span_contents),
-                        label,
-                    ))
-                }
-
-                let callable = &computation_db[user_component_id];
-                let definition_snippet =
-                    get_definition_span(callable, krate_collection, package_graph);
-                let msg = format!(
-                    "You can't return a naked generic parameter from a constructor, like `{naked_parameter}` in `{}`.\n\
-                    I don't take into account trait bounds when building your dependency graph. A constructor \
-                    that returns a naked generic parameter is equivalent, in my eyes, to a constructor that can build \
-                    **any** type, which is unlikely to be what you want!",
-                    callable.path
-                );
-                let error = anyhow::anyhow!(e).context(msg);
-                CompilerDiagnostic::builder(source, error)
-                    .optional_label(label)
-                    .optional_additional_annotated_snippet(definition_snippet)
-                    .help(
-                        "Can you return a concrete type as output? \n\
-                        Or wrap the generic parameter in a non-generic container? \
-                        For example, `T` in `Vec<T>` is not considered to be a naked parameter."
-                            .into(),
-                    )
-                    .build()
-            }
-        };
-        diagnostics.push(diagnostic.into());
-    }
-
-    fn invalid_request_handler(
-        e: RequestHandlerValidationError,
-        user_component_id: UserComponentId,
-        user_component_db: &UserComponentDb,
-        computation_db: &ComputationDb,
-        krate_collection: &CrateCollection,
-        package_graph: &PackageGraph,
-        diagnostics: &mut Vec<miette::Error>,
-    ) {
-        let location = user_component_db.get_location(user_component_id);
-        let source = match location.source_file(package_graph) {
-            Ok(s) => s,
-            Err(e) => {
-                diagnostics.push(e.into());
-                return;
-            }
-        };
-        let label = diagnostic::get_f_macro_invocation_span(&source, location)
-            .map(|s| s.labeled("The request handler was registered here".into()));
-        let diagnostic = match e {
-            RequestHandlerValidationError::CannotReturnTheUnitType
-            | RequestHandlerValidationError::CannotFalliblyReturnTheUnitType => {
-                CompilerDiagnostic::builder(source, e)
-                    .optional_label(label)
-                    .build()
-            }
-            RequestHandlerValidationError::UnderconstrainedGenericParameters { ref parameters } => {
-                fn get_definition_span(
-                    callable: &Callable,
-                    free_parameters: &IndexSet<String>,
-                    krate_collection: &CrateCollection,
-                    package_graph: &PackageGraph,
-                ) -> Option<AnnotatedSnippet> {
-                    let global_item_id = callable.source_coordinates.as_ref()?;
-                    let item = krate_collection.get_type_by_global_type_id(global_item_id);
-                    let definition_span = item.span.as_ref()?;
-                    let source_contents = diagnostic::read_source_file(
-                        &definition_span.filename,
-                        &package_graph.workspace(),
-                    )
-                    .ok()?;
-                    let span = convert_rustdoc_span(&source_contents, definition_span.to_owned());
-                    let span_contents =
-                        source_contents[span.offset()..(span.offset() + span.len())].to_string();
-                    let generic_params = match &item.inner {
-                        ItemEnum::Function(_) => {
-                            if let Ok(item) = syn::parse_str::<syn::ItemFn>(&span_contents) {
-                                item.sig.generics.params
-                            } else if let Ok(item) =
-                                syn::parse_str::<syn::ImplItemFn>(&span_contents)
-                            {
-                                item.sig.generics.params
-                            } else {
-                                panic!("Could not parse as a function or method:\n{span_contents}")
-                            }
-                        }
-                        _ => unreachable!(),
-                    };
-
-                    let mut labels = vec![];
-                    for param in generic_params {
-                        if let syn::GenericParam::Type(ty) = param {
-                            if free_parameters.contains(ty.ident.to_string().as_str()) {
-                                labels.push(
-                                    convert_proc_macro_span(&span_contents, ty.span()).labeled(
-                                        "The generic parameter without a concrete type".into(),
-                                    ),
-                                );
-                            }
-                        }
-                    }
-                    let source_path = definition_span.filename.to_str().unwrap();
-                    Some(AnnotatedSnippet::new_with_labels(
-                        NamedSource::new(source_path, span_contents),
-                        labels,
-                    ))
-                }
-
-                let callable = &computation_db[user_component_id];
-                let definition_snippet =
-                    get_definition_span(callable, parameters, krate_collection, package_graph);
-                let free_parameters = if parameters.len() == 1 {
-                    format!("`{}`", &parameters[0])
-                } else {
-                    let mut buffer = String::new();
-                    comma_separated_list(
-                        &mut buffer,
-                        parameters.iter(),
-                        |p| format!("`{}`", p),
-                        "and",
-                    )
-                    .unwrap();
-                    buffer
-                };
-                let verb = if parameters.len() == 1 { "does" } else { "do" };
-                let plural = if parameters.len() == 1 { "" } else { "s" };
-                let error = anyhow::anyhow!(e)
-                    .context(
-                        format!(
-                            "I am not smart enough to figure out the concrete type for all the generic parameters in `{}`.\n\
-                            There should no unassigned generic parameters in request handlers, but {free_parameters} {verb} \
-                            not seem to have been assigned a concrete type.",
-                            callable.path));
-                CompilerDiagnostic::builder(source, error)
-                    .optional_label(label)
-                    .optional_additional_annotated_snippet(definition_snippet)
-                    .help(
-                        format!("Specify the concrete type{plural} for {free_parameters} when registering the request handler against the blueprint: \n\
-                        |  bp.route(\n\
-                        |    ..\n\
-                        |    f!(my_crate::my_handler::<ConcreteType>), \n\
-                        |  )"))
-                    // ^ TODO: add a proper code snippet here, using the actual function that needs
-                    //    to be amended instead of a made signature
-                    .build()
-            }
-        };
-        diagnostics.push(diagnostic.into());
-    }
-
-    fn invalid_wrapping_middleware(
-        e: WrappingMiddlewareValidationError,
-        user_component_id: UserComponentId,
-        user_component_db: &UserComponentDb,
-        computation_db: &ComputationDb,
-        krate_collection: &CrateCollection,
-        package_graph: &PackageGraph,
-        diagnostics: &mut Vec<miette::Error>,
-    ) {
-        use WrappingMiddlewareValidationError::*;
-
-        let location = user_component_db.get_location(user_component_id);
-        let source = match location.source_file(package_graph) {
-            Ok(s) => s,
-            Err(e) => {
-                diagnostics.push(e.into());
-                return;
-            }
-        };
-        let label = diagnostic::get_f_macro_invocation_span(&source, location)
-            .map(|s| s.labeled("The wrapping middleware was registered here".into()));
-        let diagnostic = match e {
-            CannotReturnTheUnitType
-            | CannotFalliblyReturnTheUnitType
-            | MustTakeNextAsInputParameter => CompilerDiagnostic::builder(source, e)
-                .optional_label(label)
-                .build(),
-            CannotTakeMoreThanOneNextAsInputParameter => CompilerDiagnostic::builder(source, e)
-                .optional_label(label)
-                .help("Remove the extra `Next` input parameters until only one is left.".into())
-                .build(),
-            NextGenericParameterMustBeNaked { ref parameter } => {
-                let help =
-                    format!("Take `Next<T>` rather than `Next<{parameter}>` as input parameter in your middleware.");
-                CompilerDiagnostic::builder(source, e)
-                    .optional_label(label)
-                    .help(help)
-                    .build()
-            }
-            UnderconstrainedGenericParameters { ref parameters } => {
-                fn get_definition_span(
-                    callable: &Callable,
-                    free_parameters: &IndexSet<String>,
-                    krate_collection: &CrateCollection,
-                    package_graph: &PackageGraph,
-                ) -> Option<AnnotatedSnippet> {
-                    let global_item_id = callable.source_coordinates.as_ref()?;
-                    let item = krate_collection.get_type_by_global_type_id(global_item_id);
-                    let definition_span = item.span.as_ref()?;
-                    let source_contents = diagnostic::read_source_file(
-                        &definition_span.filename,
-                        &package_graph.workspace(),
-                    )
-                    .ok()?;
-                    let span = convert_rustdoc_span(&source_contents, definition_span.to_owned());
-                    let span_contents =
-                        source_contents[span.offset()..(span.offset() + span.len())].to_string();
-                    let generic_params = match &item.inner {
-                        ItemEnum::Function(_) => {
-                            if let Ok(item) = syn::parse_str::<syn::ItemFn>(&span_contents) {
-                                item.sig.generics.params
-                            } else if let Ok(item) =
-                                syn::parse_str::<syn::ImplItemFn>(&span_contents)
-                            {
-                                item.sig.generics.params
-                            } else {
-                                panic!("Could not parse as a function or method:\n{span_contents}")
-                            }
-                        }
-                        _ => unreachable!(),
-                    };
-
-                    let mut labels = vec![];
-                    for param in generic_params {
-                        if let syn::GenericParam::Type(ty) = param {
-                            if free_parameters.contains(ty.ident.to_string().as_str()) {
-                                labels.push(
-                                    convert_proc_macro_span(&span_contents, ty.span()).labeled(
-                                        "The generic parameter without a concrete type".into(),
-                                    ),
-                                );
-                            }
-                        }
-                    }
-                    let source_path = definition_span.filename.to_str().unwrap();
-                    Some(AnnotatedSnippet::new_with_labels(
-                        NamedSource::new(source_path, span_contents),
-                        labels,
-                    ))
-                }
-
-                let callable = &computation_db[user_component_id];
-                let definition_snippet =
-                    get_definition_span(callable, parameters, krate_collection, package_graph);
-                let free_parameters = if parameters.len() == 1 {
-                    format!("`{}`", &parameters[0])
-                } else {
-                    let mut buffer = String::new();
-                    comma_separated_list(
-                        &mut buffer,
-                        parameters.iter(),
-                        |p| format!("`{}`", p),
-                        "and",
-                    )
-                    .unwrap();
-                    buffer
-                };
-                let verb = if parameters.len() == 1 { "does" } else { "do" };
-                let plural = if parameters.len() == 1 { "" } else { "s" };
-                let error = anyhow::anyhow!(e)
-                    .context(
-                        format!(
-                            "I am not smart enough to figure out the concrete type for all the generic parameters in `{}`.\n\
-                            There should no unassigned generic parameters in wrapping middlewares apart from the one in `Next<_>`, but {free_parameters} {verb} \
-                            not seem to have been assigned a concrete type.",
-                            callable.path));
-                CompilerDiagnostic::builder(source, error)
-                    .optional_label(label)
-                    .optional_additional_annotated_snippet(definition_snippet)
-                    .help(
-                        format!("Specify the concrete type{plural} for {free_parameters} when registering the wrapping middleware against the blueprint: \n\
-                        |  bp.wrap(\n\
-                        |    f!(my_crate::my_middleware::<ConcreteType>), \n\
-                        |  )"))
-                    // ^ TODO: add a proper code snippet here, using the actual function that needs
-                    //    to be amended instead of a made signature
-                    .build()
-            }
-        };
-        diagnostics.push(diagnostic.into());
-    }
-
-    fn invalid_response_type(
-        e: MissingTraitImplementationError,
-        output_type: &ResolvedType,
-        user_component_id: UserComponentId,
-        user_component_db: &UserComponentDb,
-        package_graph: &PackageGraph,
-        diagnostics: &mut Vec<miette::Error>,
-    ) {
-        let location = user_component_db.get_location(user_component_id);
-        let raw_user_component = &user_component_db[user_component_id];
-        let callable_type = raw_user_component.callable_type();
-        let source = match location.source_file(package_graph) {
-            Ok(s) => s,
-            Err(e) => {
-                diagnostics.push(e.into());
-                return;
-            }
-        };
-        let label = diagnostic::get_f_macro_invocation_span(&source, location)
-            .map(|s| s.labeled(format!("The {callable_type} was registered here")));
-        let error = anyhow::Error::from(e).context(format!(
-            "I can't use the type returned by this {callable_type} to create an HTTP \
-                response.\n\
-                It doesn't implement `pavex::response::IntoResponse`."
-        ));
-        let help = format!("Implement `pavex::response::IntoResponse` for `{output_type:?}`.");
-        let diagnostic = CompilerDiagnostic::builder(source, error)
-            .optional_label(label)
-            .help(help)
-            .build();
-        diagnostics.push(diagnostic.into());
-    }
-
-    fn cannot_handle_into_response_implementation(
-        e: CallableResolutionError,
-        output_type: &ResolvedType,
-        raw_user_component_id: UserComponentId,
-        raw_user_component_db: &UserComponentDb,
-        package_graph: &PackageGraph,
-        diagnostics: &mut Vec<miette::Error>,
-    ) {
-        let location = raw_user_component_db.get_location(raw_user_component_id);
-        let raw_user_component = &raw_user_component_db[raw_user_component_id];
-        let callable_type = raw_user_component.callable_type();
-        let source = match location.source_file(package_graph) {
-            Ok(s) => s,
-            Err(e) => {
-                diagnostics.push(e.into());
-                return;
-            }
-        };
-        let label = diagnostic::get_f_macro_invocation_span(&source, location)
-            .map(|s| s.labeled(format!("The {callable_type} was registered here")));
-        let error = anyhow::Error::from(e).context(format!(
-            "Something went wrong when I tried to analyze the implementation of \
-                `pavex::response::IntoResponse` for {output_type:?}, the type returned by 
-                one of your {callable_type}s.\n\
-                This is definitely a bug, I am sorry! Please file an issue on \
-                https://github.com/LukeMathWalker/pavex"
-        ));
-        let diagnostic = CompilerDiagnostic::builder(source, error)
-            .optional_label(label)
-            .build();
-        diagnostics.push(diagnostic.into());
-    }
-
-    fn invalid_error_observer(
-        e: ErrorObserverValidationError,
-        raw_user_component_id: UserComponentId,
-        raw_user_component_db: &UserComponentDb,
-        computation_db: &ComputationDb,
-        krate_collection: &CrateCollection,
-        package_graph: &PackageGraph,
-        diagnostics: &mut Vec<miette::Error>,
-    ) {
-        let location = raw_user_component_db.get_location(raw_user_component_id);
-        let source = match location.source_file(package_graph) {
-            Ok(s) => s,
-            Err(e) => {
-                diagnostics.push(e.into());
-                return;
-            }
-        };
-        let label = diagnostic::get_f_macro_invocation_span(&source, location)
-            .map(|s| s.labeled("The error observer was registered here".into()));
-        let diagnostic = match &e {
-            // TODO: Add a sub-diagnostic showing the error handler signature, highlighting with
-            //  a label the non-unit return type.
-            ErrorObserverValidationError::MustReturnUnitType { .. } |
-            // TODO: Add a sub-diagnostic showing the error handler signature, highlighting with
-            //  a label the input types. 
-            ErrorObserverValidationError::DoesNotTakeErrorReferenceAsInput { .. } => {
-                CompilerDiagnostic::builder(source, e)
-                    .optional_label(label)
-                    .build()
-            }
-            ErrorObserverValidationError::UnassignedGenericParameters { ref parameters, .. } => {
-                fn get_definition_span(
-                    callable: &Callable,
-                    free_parameters: &IndexSet<String>,
-                    krate_collection: &CrateCollection,
-                    package_graph: &PackageGraph,
-                ) -> Option<AnnotatedSnippet> {
-                    let global_item_id = callable.source_coordinates.as_ref()?;
-                    let item = krate_collection.get_type_by_global_type_id(global_item_id);
-                    let definition_span = item.span.as_ref()?;
-                    let source_contents = diagnostic::read_source_file(
-                        &definition_span.filename,
-                        &package_graph.workspace(),
-                    )
-                        .ok()?;
-                    let span = convert_rustdoc_span(&source_contents, definition_span.to_owned());
-                    let span_contents =
-                        source_contents[span.offset()..(span.offset() + span.len())].to_string();
-                    let generic_params = match &item.inner {
-                        ItemEnum::Function(_) => {
-                            if let Ok(item) = syn::parse_str::<syn::ItemFn>(&span_contents) {
-                                item.sig.generics.params
-                            } else if let Ok(item) =
-                                syn::parse_str::<syn::ImplItemFn>(&span_contents)
-                            {
-                                item.sig.generics.params
-                            } else {
-                                panic!("Could not parse as a function or method:\n{span_contents}")
-                            }
-                        }
-                        _ => unreachable!(),
-                    };
-
-                    let mut labels = vec![];
-                    for param in generic_params {
-                        if let syn::GenericParam::Type(ty) = param {
-                            if free_parameters.contains(ty.ident.to_string().as_str()) {
-                                labels.push(
-                                    convert_proc_macro_span(&span_contents, ty.span())
-                                        .labeled("I can't infer this".into()),
-                                );
-                            }
-                        }
-                    }
-                    let source_path = definition_span.filename.to_str().unwrap();
-                    Some(AnnotatedSnippet::new_with_labels(
-                        NamedSource::new(source_path, span_contents),
-                        labels,
-                    ))
-                }
-
-                let callable = &computation_db[raw_user_component_id];
-                let definition_snippet =
-                    get_definition_span(callable, parameters, krate_collection, package_graph);
-                CompilerDiagnostic::builder(source, e)
-                    .optional_label(label)
-                    .optional_additional_annotated_snippet(definition_snippet)
-                    .help(
-                        "Specify the concrete type(s) for the problematic \
-                        generic parameter(s) when registering the error observer against the blueprint: `f!(my_crate::my_observer::<ConcreteType>)`".into())
-                    // ^ TODO: add a proper code snippet here, using the actual function that needs
-                    //    to be amended instead of a made signature
-                    .build()
-            }
-        };
-        diagnostics.push(diagnostic.into());
-    }
-
-    fn invalid_error_handler(
-        e: ErrorHandlerValidationError,
-        raw_user_component_id: UserComponentId,
-        raw_user_component_db: &UserComponentDb,
-        computation_db: &ComputationDb,
-        krate_collection: &CrateCollection,
-        package_graph: &PackageGraph,
-        diagnostics: &mut Vec<miette::Error>,
-    ) {
-        let location = raw_user_component_db.get_location(raw_user_component_id);
-        let source = match location.source_file(package_graph) {
-            Ok(s) => s,
-            Err(e) => {
-                diagnostics.push(e.into());
-                return;
-            }
-        };
-        let label = diagnostic::get_f_macro_invocation_span(&source, location)
-            .map(|s| s.labeled("The error handler was registered here".into()));
-        let diagnostic = match &e {
-            // TODO: Add a sub-diagnostic showing the error handler signature, highlighting with
-            // a label the missing return type.
-            ErrorHandlerValidationError::CannotReturnTheUnitType(_) |
-            // TODO: Add a sub-diagnostic showing the error handler signature, highlighting with
-            // a label the input types. Perhaps add a signature showing the signature of
-            // the associate fallible handler, highlighting the output type.
-            ErrorHandlerValidationError::DoesNotTakeErrorReferenceAsInput { .. } => {
-                CompilerDiagnostic::builder(source, e)
-                    .optional_label(label)
-                    .build()
-            }
-            ErrorHandlerValidationError::UnderconstrainedGenericParameters { ref parameters, ref error_ref_input_index } => {
-                fn get_definition_span(
-                    callable: &Callable,
-                    free_parameters: &IndexSet<String>,
-                    error_ref_input_index: usize,
-                    krate_collection: &CrateCollection,
-                    package_graph: &PackageGraph,
-                ) -> Option<AnnotatedSnippet> {
-                    let global_item_id = callable.source_coordinates.as_ref()?;
-                    let item = krate_collection.get_type_by_global_type_id(global_item_id);
-                    let definition_span = item.span.as_ref()?;
-                    let source_contents = diagnostic::read_source_file(
-                        &definition_span.filename,
-                        &package_graph.workspace(),
-                    )
-                        .ok()?;
-                    let span = convert_rustdoc_span(&source_contents, definition_span.to_owned());
-                    let span_contents =
-                        source_contents[span.offset()..(span.offset() + span.len())].to_string();
-                    let (generic_params, error_input) = match &item.inner {
-                        ItemEnum::Function(_) => {
-                            if let Ok(item) = syn::parse_str::<syn::ItemFn>(&span_contents) {
-                                (item.sig.generics.params, item.sig.inputs[error_ref_input_index].clone())
-                            } else if let Ok(item) =
-                                syn::parse_str::<syn::ImplItemFn>(&span_contents)
-                            {
-                                (item.sig.generics.params, item.sig.inputs[error_ref_input_index].clone())
-                            } else {
-                                panic!("Could not parse as a function or method:\n{span_contents}")
-                            }
-                        }
-                        _ => unreachable!(),
-                    };
-
-                    let mut labels = vec![];
-                    let subject_verb = if generic_params.len() == 1 {
-                        "it is"
-                    } else {
-                        "they are"
-                    };
-                    for param in generic_params {
-                        if let syn::GenericParam::Type(ty) = param {
-                            if free_parameters.contains(ty.ident.to_string().as_str()) {
-                                labels.push(
-                                    convert_proc_macro_span(&span_contents, ty.span())
-                                        .labeled("I can't infer this..".into()),
-                                );
-                            }
-                        }
-                    }
-                    let error_input_span = error_input.span();
-                    labels.push(
-                        convert_proc_macro_span(&span_contents, error_input_span)
-                            .labeled(format!("..because {subject_verb} not used here")),
-                    );
-                    let source_path = definition_span.filename.to_str().unwrap();
-                    Some(AnnotatedSnippet::new_with_labels(
-                        NamedSource::new(source_path, span_contents),
-                        labels,
-                    ))
-                }
-
-                let callable = &computation_db[raw_user_component_id];
-                let definition_snippet =
-                    get_definition_span(callable, parameters, *error_ref_input_index, krate_collection, package_graph);
-                let subject_verb = if parameters.len() == 1 {
-                    "it isn't"
-                } else {
-                    "they aren't"
-                };
-                let free_parameters = if parameters.len() == 1 {
-                    format!("`{}`", &parameters[0])
-                } else {
-                    let mut buffer = String::new();
-                    comma_separated_list(&mut buffer, parameters.iter(), |p| format!("`{}`", p), "and").unwrap();
-                    buffer
-                };
-                let error = anyhow::anyhow!(e)
-                    .context(
-                        format!(
-                            "I am not smart enough to figure out the concrete type for all the generic parameters in `{}`.\n\
-                            I can only infer the type of an unassigned generic parameter if it appears in the error type processed by this error handler. This is \
-                            not the case for {free_parameters}, since {subject_verb} used by the error type.",
-                            callable.path));
-                CompilerDiagnostic::builder(source, error)
-                    .optional_label(label)
-                    .optional_additional_annotated_snippet(definition_snippet)
-                    .help(
-                        "Specify the concrete type(s) for the problematic \
-                        generic parameter(s) when registering the error handler against the blueprint: \n\
-                        |  .error_handler(\n\
-                        |    f!(my_crate::my_error_handler::<ConcreteType>)\n\
-                        |  )".into())
-                    // ^ TODO: add a proper code snippet here, using the actual function that needs
-                    //    to be amended instead of a made signature
-                    .build()
-            }
-        };
-        diagnostics.push(diagnostic.into());
-    }
-
-    fn error_handler_for_infallible_component(
-        error_handler_id: UserComponentId,
-        fallible_id: UserComponentId,
-        raw_user_component_db: &UserComponentDb,
-        package_graph: &PackageGraph,
-        diagnostics: &mut Vec<miette::Error>,
-    ) {
-        let fallible_kind = raw_user_component_db[fallible_id].callable_type();
-        let location = raw_user_component_db.get_location(error_handler_id);
-        let source = match location.source_file(package_graph) {
-            Ok(s) => s,
-            Err(e) => {
-                diagnostics.push(e.into());
-                return;
-            }
-        };
-        let label = diagnostic::get_f_macro_invocation_span(&source, location)
-            .map(|s| s.labeled("The unnecessary error handler was registered here".into()));
-        let error = anyhow::anyhow!(
-            "You registered an error handler for a {} that doesn't return a `Result`.",
-            fallible_kind
-        );
-        let diagnostic = CompilerDiagnostic::builder(source, error)
-            .optional_label(label)
-            .help(format!(
-                "Remove the error handler, it is not needed. The {fallible_kind} is infallible!"
-            ))
-            .build();
-        diagnostics.push(diagnostic.into());
-    }
-
-    fn error_handler_for_a_singleton(
-        error_handler_id: UserComponentId,
-        fallible_id: UserComponentId,
-        raw_user_component_db: &UserComponentDb,
-        package_graph: &PackageGraph,
-        diagnostics: &mut Vec<miette::Error>,
-    ) {
-        debug_assert_eq!(
-            raw_user_component_db[fallible_id].callable_type(),
-            CallableType::Constructor
-        );
-        let location = raw_user_component_db.get_location(error_handler_id);
-        let source = match location.source_file(package_graph) {
-            Ok(s) => s,
-            Err(e) => {
-                diagnostics.push(e.into());
-                return;
-            }
-        };
-        let label = diagnostic::get_f_macro_invocation_span(&source, location)
-            .map(|s| s.labeled("The unnecessary error handler was registered here".into()));
-        let error = anyhow::anyhow!(
-            "You can't register an error handler for a singleton constructor. \n\
-                If I fail to build a singleton, I bubble up the error - it doesn't get handled.",
-        );
-        let diagnostic = CompilerDiagnostic::builder(source, error)
-            .optional_label(label)
-            .help("Remove the error handler, it is not needed.".to_string())
-            .build();
-        diagnostics.push(diagnostic.into());
-    }
-
-    fn missing_error_handler(
-        fallible_id: UserComponentId,
-        raw_user_component_db: &UserComponentDb,
-        package_graph: &PackageGraph,
-        diagnostics: &mut Vec<miette::Error>,
-    ) {
-        let fallible_kind = raw_user_component_db[fallible_id].callable_type();
-        let location = raw_user_component_db.get_location(fallible_id);
-        let source = match location.source_file(package_graph) {
-            Ok(s) => s,
-            Err(e) => {
-                diagnostics.push(e.into());
-                return;
-            }
-        };
-        let label = diagnostic::get_f_macro_invocation_span(&source, location)
-            .map(|s| s.labeled(format!("The fallible {fallible_kind} was registered here")));
-        let error = anyhow::anyhow!(
-                "You registered a {fallible_kind} that returns a `Result`, but you did not register an \
-                 error handler for it. \
-                 If I don't have an error handler, I don't know what to do with the error when the \
-                 {fallible_kind} fails!",
-            );
-        let diagnostic = CompilerDiagnostic::builder(source, error)
-            .optional_label(label)
-            .help("Add an error handler via `.error_handler`".to_string())
-            .build();
-        diagnostics.push(diagnostic.into());
     }
 }
